Add basic postfix mail config in use on foghlaim.com
At some point it's best if this is transferred to a repository that manages the server, but for now it will do as part of this repository.

Just a basic config at the moment to get email sending from @foghlaim.com. Hopefully I can tweak this out and make it smoother all around, possibly for multiple domains.