Hide inactive Jetpack modules, ahhhh
I had so many deactivated Jetpack modules in my list of Jetpack modules that it was really hard to tell what was going on. This plugin hides all the inactive modules so that I can see what things really look like.

Of course, I'll have to disable the plugin any time I want to add another module, but that's ok with me. :)

Actually, I'll have to watch for the next Jetpack update. It's possible that this can display new modules if they are marked somehow even if they are inactive.